# [Core] Performance fixes in GameProfilingSystem and ChromeTracingProfileWriter