--- a/sources/core/Stride.Core/Diagnostics/ChromeTracingProfileWriter.cs
+++ b/sources/core/Stride.Core/Diagnostics/ChromeTracingProfileWriter.cs
@@ -6,6 +6,8 @@ using System.Text.Json.Nodes;
 using System.Text.Json;
 using System.Threading.Channels;
 using System.Threading.Tasks;
+using System.Threading;
+using System;
 
 namespace Stride.Core.Diagnostics
 {
@@ -23,6 +25,7 @@ namespace Stride.Core.Diagnostics
         public void Start(string outputPath, bool indentOutput = false)
         {
             eventReader = Profiler.Subscribe();
+            cts = new CancellationTokenSource();
             writerTask = Task.Run(async () =>
             {
                 var pid = Process.GetCurrentProcess().Id;
@@ -54,43 +57,47 @@ namespace Stride.Core.Diagnostics
                 writer.WriteEndObject();
                 writer.WriteEndObject();
 
-                await foreach (var e in eventReader.ReadAllAsync())
+                try
                 {
-                    //gc scopes currently start at negative timestamps and should be filtered out,
-                    //because they don't represent durations.
-                    if (e.TimeStamp.Ticks < 0)
-                        continue;
-
-                    double startTimeInMicroseconds = e.TimeStamp.TotalMilliseconds * 1000.0;
-                    double durationInMicroseconds = e.ElapsedTime.TotalMilliseconds * 1000.0;
-
-                    Debug.Assert(durationInMicroseconds >= 0);
-
-                    writer.WriteStartObject();
-                    writer.WriteString("name", e.Key.Name);
-                    if (e.Key.Parent != null)
-                        writer.WriteString("cat", e.Key.Parent.Name);
-                    writer.WriteString("ph", "X");
-                    writer.WriteNumber("ts", startTimeInMicroseconds);
-                    writer.WriteNumber("dur", durationInMicroseconds);
-                    writer.WriteNumber("tid", e.ThreadId >= 0 ? e.ThreadId : int.MaxValue);
-                    writer.WriteNumber("pid", pid);
-                    if (e.Attributes.Count > 0)
+                    await foreach (var e in eventReader.ReadAllAsync(cts.Token))
                     {
-                        writer.WriteStartObject("args");
-                        foreach (var (k, v) in e.Attributes)
+                        //gc scopes currently start at negative timestamps and should be filtered out,
+                        //because they don't represent durations.
+                        if (e.TimeStamp.Ticks < 0)
+                            continue;
+
+                        double startTimeInMicroseconds = e.TimeStamp.TotalMilliseconds * 1000.0;
+                        double durationInMicroseconds = e.ElapsedTime.TotalMilliseconds * 1000.0;
+
+                        Debug.Assert(durationInMicroseconds >= 0);
+
+                        writer.WriteStartObject();
+                        writer.WriteString("name", e.Key.Name);
+                        if (e.Key.Parent != null)
+                            writer.WriteString("cat", e.Key.Parent.Name);
+                        writer.WriteString("ph", "X");
+                        writer.WriteNumber("ts", startTimeInMicroseconds);
+                        writer.WriteNumber("dur", durationInMicroseconds);
+                        writer.WriteNumber("tid", e.ThreadId >= 0 ? e.ThreadId : int.MaxValue);
+                        writer.WriteNumber("pid", pid);
+                        if (e.Attributes.Count > 0)
                         {
-                            writer.WriteString(k, v.ToString());
+                            writer.WriteStartObject("args");
+                            foreach (var (k, v) in e.Attributes)
+                            {
+                                writer.WriteString(k, v.ToString());
+                            }
+                            writer.WriteEndObject();
                         }
                         writer.WriteEndObject();
-                    }
-                    writer.WriteEndObject();
 
-                    if (writer.BytesPending >= 1024 * 1024)
-                    {
-                        await writer.FlushAsync();
+                        if (writer.BytesPending >= 1024 * 1024)
+                        {
+                            await writer.FlushAsync();
+                        }
                     }
                 }
+                catch (OperationCanceledException) { } // cancellation was requested, let's finish
 
                 writer.WriteEndArray();
                 writer.WriteEndObject();
@@ -106,6 +113,11 @@ namespace Stride.Core.Diagnostics
             if (eventReader != null)
             {
                 Profiler.Unsubscribe(eventReader);
+                eventReader = null;
+
+                cts?.Cancel();
+                cts?.Dispose();
+                
                 writerTask?.Wait();
             }
         }
@@ -113,6 +125,7 @@ namespace Stride.Core.Diagnostics
 #nullable enable
         ChannelReader<ProfilingEvent>? eventReader;
         Task? writerTask;
+        CancellationTokenSource? cts;
 #nullable disable
     }