--- a/sources/core/Stride.Core/Diagnostics/ProfilingEventMessage.cs
+++ b/sources/core/Stride.Core/Diagnostics/ProfilingEventMessage.cs
@@ -2,6 +2,7 @@
 // Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
 
 using System.Text;
+using System.Threading;
 
 namespace Stride.Core.Diagnostics
 {
@@ -10,6 +11,14 @@ namespace Stride.Core.Diagnostics
     /// </summary>
     public struct ProfilingEventMessage
     {
+        // The ProfilingCustomValue holds a struct which would need to be boxed for string formatting.
+        // To avoid this, we use a custom formatter object which can allocate statically.
+        private static readonly ThreadLocal<ProfilingCustomValueFormatter> formatter0 = new(() => new ProfilingCustomValueFormatter());
+        private static readonly ThreadLocal<ProfilingCustomValueFormatter> formatter1 = new(() => new ProfilingCustomValueFormatter());
+        private static readonly ThreadLocal<ProfilingCustomValueFormatter> formatter2 = new(() => new ProfilingCustomValueFormatter());
+        private static readonly ThreadLocal<ProfilingCustomValueFormatter> formatter3 = new(() => new ProfilingCustomValueFormatter());
+        private static readonly ThreadLocal<object[]> formatterParams = new(() => [formatter0.Value, formatter1.Value, formatter2.Value, formatter3.Value]);
+
         /// <summary>
         /// The text supporting formatting of up to 4 numerical parameters.
         /// </summary>
@@ -44,12 +53,40 @@ namespace Stride.Core.Diagnostics
 
         public override string ToString()
         {
-            return string.Format(Text, Custom0?.ToObject(), Custom1?.ToObject(), Custom2?.ToObject(), Custom3?.ToObject());
+            PopulateFormatters();
+            return string.Format(Text, formatterParams.Value);
         }
 
         public void ToString(StringBuilder builder)
         {
-            builder.AppendFormat(Text, Custom0?.ToObject(), Custom1?.ToObject(), Custom2?.ToObject(), Custom3?.ToObject());
+            PopulateFormatters();
+            builder.AppendFormat(Text, formatterParams.Value);
+        }
+
+        private void PopulateFormatters()
+        {
+            if (Custom0.HasValue)
+                formatter0.Value.Value = Custom0.Value;
+            if (Custom1.HasValue)
+                formatter1.Value.Value = Custom1.Value;
+            if (Custom2.HasValue)
+                formatter2.Value.Value = Custom2.Value;
+            if (Custom3.HasValue)
+                formatter3.Value.Value = Custom3.Value;
+        }
+
+        private class ProfilingCustomValueFormatter
+        {
+            public ProfilingCustomValue Value { get; set; }
+
+            public override string ToString()
+            {
+                if (Value.ValueType == typeof(int)) return Value.IntValue.ToString();
+                else if (Value.ValueType == typeof(float)) return Value.FloatValue.ToString();
+                else if (Value.ValueType == typeof(long)) return Value.LongValue.ToString();
+                else if (Value.ValueType == typeof(double)) return Value.DoubleValue.ToString();
+                else return "<unknown>";
+            }
         }
     }
 }