# PR Details

## Description

This is a continuation of #1953 to fix performance issues with the Profiler.
As part of this work I ended up producing #2049.

From #1953
* Fix busy-wait in `GameProfilingSystem` and make surrounding code clearer.
* Add manual flushing to the `JsonWriter`, so the internal buffer doesn't grow indefinitely and put it on the fast path by skipping validation.

Additionally:
* Refactor `GameProfilingSystem` to extract all string formatting related logic into a private class `GameProfilingSystemStrings`
* Modify a lot of formatting code to allocate less
* Interchange between two different tasks in the `GameProfilingSystem` depending on the filtering mode
* Add cancellation tokens to the Profiler subscribers
* Remove semaphore lock from the Profiler in favor of ConcurrentDictionary with a flag checked using Interlocked
* Move from Unbounded channels to Bounded to reduce allocation and incredible GC pressure
* Add Task.Delay to consumption of events in GameProfilingSystem to further reduce allocation by waiting on channels

## Related Issue

Closes #1953

## Motivation and Context

The Profiler should be very efficient. After all if by measuring performance we decrease it, it's not a great system.
These changes aim to improve mainly GC pressure by minimizing allocations.
There's still further field for improvement that needs some deeper investigation - I was aiming to achieve perf parity between Fps profiler mode and CpuEvents -> it's close but there's still a bit of a difference. Furthermore there's still difference between measuring FPS without Profiling turned on and with.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [x] Refactoring
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [ ] My change requires a change to the documentation.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.
- [ ] **I have built and run the editor to try this change out.**