@manio143 2023-11-28
I'm gonna try and create some BenchmarkDotNet samples to ensure my evidence wasn't incorrectly collected.