@Kryptos-FR 2025-02-24
@manio143 I can fix the merge conflicts introduced by #2643 if you want. I'll need to be added as a contributor to your repo in that case. Let me know.